+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `ElementIdConverter` to support Revit 2024+
 ### Fixes
 - Fix `ObjectJsonConverter` to support `null` value.
+- Fix `XYZConverter` to support `null` value.
 
 ## [1.0.3] / 2022-10-10
 ### Features